Setting release to 0.0.1

+++ pylibcontainer/utils.py
@@ -5,7 +5,7 @@ import math
 def eprint(*args, **kwargs):
     print(*args, file=sys.stderr, **kwargs)
 
-class HumanSize(long):
+class HumanSize(int):
     """ define a size class to allow custom formatting
         Implements a format specifier of S for the size class - which displays a human readable in b, kb, Mb etc
     """
@@ -13,7 +13,7 @@ class HumanSize(long):
         if fmt == "" or fmt[-1] != "S":
             if fmt[-1].tolower() in ['b', 'c', 'd', 'o', 'x', 'n', 'e', 'f', 'g', '%']:
                 # Numeric format.
-                return long(self).__format__(fmt)
+                return int(self).__format__(fmt)
             return str(self).__format__(fmt)
 
         val, text = float(self), ["b ", "Kb", "Mb", "Gb", "Tb", "Pb"]
